mux: Support settling time
Muxes never switch from one stable state to another instantaneously.
If a driver switches the mux, it needs to delay until the mux has
settled or else the mux will be in an undefined state.

The delay is specific to the hardware, so support setting a per-mux
delay in the core and amend the gpio-mux driver to allow specifying the
delay in the devicetree.

Signed-off-by: Lukas Wunner <lukas@wunner.de>