printk: Unregister boot consoles on register of 1st real console
The code comment preceding register_console() claims that:

   "There are two types of consoles - bootconsoles (early_printk) and
    "real" consoles (everything which is not a bootconsole) which are
    handled differently. [...]
    As soon as a "real" console is registered, all bootconsoles
    will be unregistered automatically."

But that's not what the code does:  The code unregisters bootconsoles
only when the *preferred* console registers, i.e. the last one on the
command line.  If that console's driver never registers (e.g. because
it is disabled in the kernel config), bootconsoles stay around
indefinitely.  Should the command line contain both a bootconsole as
well as a real console on a particular serial port, all messages are
logged twice once the real console registers.  Also, the log buffer is
replayed once the real console registers even though the messages were
already emitted by the bootconsole.

Amend the code to be congruent with the above-quoted code comment to
avoid these issues.

Signed-off-by: Lukas Wunner <lukas@wunner.de>
[p.rosenberger: fixup to 5.10]
Signed-off-by: Philipp Rosenberger <p.rosenberger@kunbus.com>