dts/revpi-connect: Add KSZ8851 reset delay workaround
The KSZ8851 on the RevPi Connect has a circuit which will keep it for
around 80ms in the reset when the reset gpio is released. The driver
needs to wait this time after the reset gpio is released. The driver has
no support for this. So we use a regulator with the approprieate delays
as workaround.

The startup-delay is set to 80ms. So The driver should wait for at least
80ms after enabling the regulator. The off-on-delay is set to 11.5ms.
The data sheet of the KSZ8851 says it needs a 10ms reset and the circuit
adds 1.5ms to drain the capacitor.

Signed-off-by: Philipp Rosenberger <p.rosenberger@kunbus.com>