dts/connect: Fix pibridge rs485 pulls
The TX pin doesn't need any pull as it is driven by the controller.
It makes more sense to have a pull-up at the RX pin, as the default on
the line should be 3v3.
The RTS pin is again an output pin and doesn't need any pull.
This also applies to the TERM gpio.

Also set status = "okay" in the rs485 node to make clear the uart0 is
enabled.

Signed-off-by: Philipp Rosenberger <p.rosenberger@kunbus.com>