dts/core: Use own fragments to modify existing pinmuxes
If we redefine muxes which are already defined in the devicetree we
might get errors when loading the overlay at runtime. There is no
usefull error message. It just says the loding of the overlay faild. The
bootloader seems not so picky and loads the overlay.

We move each pinmux which is already defined in bcm2708-rpi-cm.dts,
bcm2710-rpi-cm3.dts or bcm2711-rpi-cm4s.dts, to its own fragment. And we
modify the already defined pinmux instead of creating a new one.

Signed-off-by: Philipp Rosenberger <p.rosenberger@kunbus.com>