pibridge-serdev: remove superflous assignment in pibridge_recv_timeout()
Do not store the return value of wait_event_timeout() in function
pibridge_recv_timeout(), since the value is not evaluated anyway.
Also we can get rid of the local variable named "jiffies" which due to the
same name like the kernel global variable that tracks the number of ticks
since system startup may lead to confusion or even issues if in future code
changes the local declaration is accidentally removed.

Signed-off-by: Lino Sanfilippo <l.sanfilippo@kunbus.com>