arm/dts: Add overlay for RevPi HAT EEPROMs
This overlay can be used to make the RevPi HAT EEPROM available. The
overlay should only be used to flash the HAT EEPROM. Loading this
overlay by default might have undesired side effects. The HAT EEPROM is
attached to the i2c0 which is shared with the video core. On some RevPi
devices it might trigger a reset of the periphery if the reset is not
actively disabled. How the reset is disabled is board specific.

Signed-off-by: Philipp Rosenberger <p.rosenberger@kunbus.com>