pibridge-serdev: correct size of len parameter
There is no need to use a u16 data type for variables that hold the length
of IO or GATEWAY packets (this is since the length of an IO packet can be
at max 256 and the length of a GATEWAY at max 32 bytes).
So adjust the data types of local variables as well as variables in the
function parameter lists.

Signed-off-by: Lino Sanfilippo <l.sanfilippo@kunbus.com>