pibridge-serdev: Move definitions in new header file pibridge.h
Move structure definitions for GATE and IO packet headers as well as the
macro for the fifo size into the new header file pibridge.h.
By doing this extend the fifo size to a power of 2 since this is what the
kfifo API will extend the fifo to, anyway. Also rename the macro from
PIBRIDGE_RECV_BUFFER_SIZE to the more appropriate PIBRIDGE_RECV_FIFO_SIZE.

This is in preparation of implementing tracing macros which are defined in
a seperate header file but need access to the definitions mentioned above.

Signed-off-by: Lino Sanfilippo <l.sanfilippo@kunbus.com>