pibridge-serdev: fix bad cycle times
Use wait_event_hrtimeout() instead of wait_event_timeout() to wait for more
data. The latter uses jiffies to specify the timeout which requires a
minimum of 2 jiffies (which equals 20 msecs) to avoid a timeout before the
the IO timeout interval used by the pibridge module (10 msecs) have passed.

With wait_event_hrtimeout() it is possible to specify the exact 10 msecs
as a timeout so we do not have to wait up to 20 msecs in case that no more
data arrives.

Signed-off-by: Lino Sanfilippo <l.sanfilippo@kunbus.com>