dts: connect4: Use correct sniff pin order
The piBridge kernel driver needs to have the sniff pin definition in the
devicetree in a defined order. There are two properties, one for the
left piBridge (left-sniff-gpios) and one for the piBridge on the right
side of the device (right-sniff-gpios). Each of them awaits two GPIOs,
one of them is the SNIFF pin, the other is the ACKNOWLEDGE pin.

The correct order for the gpios is SNIFF pin first, then the ACK pin.
Sometimes SNIFF is called 1A for the left piBridge or 1B for the right
piBrdige, the ACK is sometimes seen as 2A or 2B for left or right
piBridge (which are the pin numbers and side on the connector).

In this case, the order was swapped. This commit fixes the order to
SNIFF pin first and then ACK pin for each of the piBridges.

Signed-off-by: Nicolai Buchwitz <n.buchwitz@kunbus.com>
Signed-off-by: Frank Erdrich <f.erdrich@kunbus.com>