usbnet: smsc95xx: Avoid duplicate MAC address from command line
If a MAC address is specified on the command line via "smsc95xx.macaddr"
and multiple smsc95xx network interfaces are attached, they all get the
same MAC address.

Avoid by incrementing the non-OUI portion of the MAC address with each
additional interface.

If this results in an invalid MAC address, the driver falls back to
generating a random one.

Inspired by the following commit by Philipp:
https://github.com/RevolutionPi/linux/commit/bf7a050e4b50

Philipp recalls that this feature was added at the behest of EdF.

Signed-off-by: Lukas Wunner <lukas@wunner.de>
Cc: Philipp Rosenberger <p.rosenberger@kunbus.com>