ARM: dts: revpi-connect4: Swap LED mode of SoC Ethernet
Existing RevPi products use a green LED to signify link presence and an
amber LED to signify activity.

The Connect 4's LAN7800 Ethernet retains that color scheme, but the SoC
Ethernet does not:  It uses green for activity and amber for link.

Invert the color scheme of the SoC Ethernet for consistency.

An effort is ongoing in mainline to introduce a standardized API and
DeviceTree schema for Ethernet LED modes.  It's still in review though,
so use the Foundation's custom DeviceTree property for now:

https://lore.kernel.org/all/20230427001541.18704-1-ansuelsmth@gmail.com/

Signed-off-by: Lukas Wunner <lukas@wunner.de>