package: Add linux-image-... meta package
As the kernel package contains the kernel version, a meta package is
needed to provide proper dependency. The meta package name is derived
from the source package name. A revpi-v8-linux package source name will
result in a linux-image-revpi-v8 meta package. The meta package depends
on the linux-image-${VERSION} package.

Signed-off-by: Philipp Rosenberger <p.rosenberger@kunbus.com>
[thomas: update for debian scripts from v6.6..v6.9]
Signed-off-by: Thomas Böhler <t.boehler@kunbus.com>