dts/core-2022: ks8851 reset gpio is active low
The reset gpio on the ksz8851 ethernet phy is active low. This was not noticed
with kernel 5.10, as the ks8851 driver used absolute values until commit
7b77bb5c ("net: ks8851: switch to using gpiod API").

This commit will fix the problem with RevPi Core S 1.1 and missing pileft /
piright interfaces on Kernel 6.1. On other devices this is not an issue, as
older Core and Connect are using a fixed regulator for reset.

Signed-off-by: Nicolai Buchwitz <n.buchwitz@kunbus.com>