pibridge-serdev: add error and info messages at startup
If pibridge_probe() fails for any other reason than failed memory
allocation there is no hint for this in the kernel log file (also the
caller of pibridge_probe() does not print an according error message).
So add error messages directly in this function and also print an info
message in case that probing was successful.

Signed-off-by: Lino Sanfilippo <l.sanfilippo@kunbus.com>