arm: dts: revpi-flat-s-2022: Use dwc2 usb
The XHCI usb controller can only be used if `otg_mode=1` is present in
/boot/config.txt. Although Raspberry Pi is about to release an updated
version of their bootloader which can detect if the xhci controller is
wanted from dt, we cannot use this feature without a proper release from
them. Thus use the dwc2 controller like we do on other devices for now.

Signed-off-by: Nicolai Buchwitz <n.buchwitz@kunbus.com>