rtc: pcf2127: Disable time stamping support
For the watchdog to work all interrupt flags need to be cleared. If any
interrupt flag is set the watchdog counter will stop at 1 and will not
trigger a watchdog reset.

The timestamp feature of the PCF2127/PCF2129/PCF2131 can set two
different interrupt flags. The TSF1 is set when the TS input is driven
to an intermediate level between power supply and ground. And the TSF2
is set when TS input is driven to ground.

When the main power is removed and the power switches to the battery
this bit might get set. It is unclear why this happens as neither the
condition for TSF2 or TSF1 should apply in this case. The TS has an
internal pull-up. If the TS pin in not connected should always have the
same level as the power supply. And the current on the TS pin should
never drop to ground level.

The second assumption, about the TSF2 seems to be true. But the TSF1 is
set after a power cycle (on some devices). And this prevents a reliable
use of the watchdog.

Disable the time stamping feature for now. As we don't have any device
which has the TS connected. This is needed for the Connect 4 watchdog to
work properly.

Signed-off-by: Philipp Rosenberger <p.rosenberger@kunbus.com>