serial: pl011: handle RS485 tx path in worker thread
For RS485 handle the tx path in a worker thread instead in the interrupt
handler. For this do not call pl011_tx_chars() from the interrupt handler
but only queue the work which then executes pl011_tx_chars().
This is the first step to avoid hugging the CPU for too long, since in
contrary to the interrupt handler the worker queue is interruptible and
may sleep.

Signed-off-by: Lino Sanfilippo <l.sanfilippo@kunbus.com>