serial: pl011: temporarily release the ports lock for delays
There are some delays needed on the tx path when RS485 is enabled:
1. In case that delays are specified before or after RTS is enabled
2. For the draining interval after the transmission of a data chunk has
   completed

In both cases the CPU is hugged for the complete delay period, since
holding the ports lock disables preemption. Improve the situation by
temporarily releasing the lock before the delay and reclaiming it
afterwards.

This change reduces almost completely all receive errors on the piBridge
which result from hugging the CPU for too long when data is sent on a
concurrent RS485 connection.

Signed-off-by: Lino Sanfilippo <l.sanfilippo@kunbus.com>