package: Add gen_revpi_pkg script for deb
The script "gen_revpi_pkg" takes the upstream version in the form
"v${MAINLINE_VER}-rt${RT_VER}-revpi${REVPI_VER}" and creates the
debianisation for the kernel. This can and should be done on every new
release which is possible because the script is mostly idempotent.

By default, the email and full name of the maintainer for the created
debianisation is "KUNBUS Support <support@kunbus.com>", the same as all
other Revolution Pi packages.

Packages are only created for arm64 as, starting with our bookworm
image, which Revolution Pi's linux kernel v6.6 will be released
alongside, armhf support is dropped, and no other architecture is
supported by us. This is also why the architecture suffix of the
packages is "-v8".

Signed-off-by: Thomas Böhler <t.boehler@kunbus.com>