arm: dts: revpi-flat-s-2022: Add overlay for dsa
Add separate overlay for enabling Flat S dsa switch functionality. There
is already a dtparam override, but this method is error-prone as it
requires the exact placement of the dtparam value (needs to be first
element before any additional dtoverlay specs). Having a distinct
overlay for this make it easier for customers to enable the feature and
this can be also incorporated into our tooling to avoid editing the
config.txt manually.

Signed-off-by: Nicolai Buchwitz <n.buchwitz@kunbus.com>