package: debian: ignore package-has-long-file-name with lintian
The resulting file names of the linux-headers-* and linux-image-*
packages are very long and lintian doesn't like this. These can't be
shortened so ignore the lintian warning altogether.

For lintian to be able to ignore the warnings the override files have to
be installed into the system. The debhelper "dh_lintian" is introduced
for this purpose.

Signed-off-by: Thomas Böhler <t.boehler@kunbus.com>