arm/dts: Add SPDX license header GPL-2.0-only
The RevPi devicetree overlays are missing license headers. Add the
appropriate SPDX license identifier and file copyright headers.

Some have a non-SPDX copyright header. Change them to SPDX as well.

One file uses GPL-2.0-or-later change to GPL-2.0-only to comply the
rest of the kernel.

One file has GPL-2.0 as license identifier, change this to GPL-2.0-only
to make this a a valid SPDX identifier as well.

The blob files are excluded, as these are mostly copies of the files
provided by Raspberry Pi with only trivial changes.

Signed-off-by: Philipp Rosenberger <p.rosenberger@kunbus.com>