package: mkdebian: force copyright generation
The debian/copyright file is copied from
scripts/package/debian/copyright and represents a generic copyright form
that is not in the standard DEP-5 form.

It should be possible to keep a debian/copyright while regenerating the
debian/ folder to have a custom debian/copyright that isn't removed.

The default debian/copyright file from scripts/package/debian/copyright
is used if the environment variable "FORCE_COPYRIGHT_GEN" is set.
No debian/copyright file is created by default and an error will be
thrown if at the end of the mkdebian script no debian/copyright file
exists.

Suggested-by: Philipp Rosenberger <p.rosenberger@kunbus.com>
Signed-off-by: Thomas Böhler <t.boehler@kunbus.com>