package: propagate necessary env variables for debuild
The kernel should be built with the environment variables
"KBUILD_BUILD_USER" and "KBUILD_BUILD_HOST" set correctly. Additionally,
the variable "KBUILD_BUILD_TIMESTAMP" should be set to ensure the builds
are reproducible.
If the kernel is built with "git-buildpackage" these variables are not
passed to Kbuild because "git-buildpackage" is calling "debuild", the
program that calls "dpkg-buildpackage" to actually build the package.
"debuild" sanitises the environment variables before calling
"dpkg-buildpackage", cleaning up the "KBUILD_BUILD_*" variables in the
process. This makes Kbuild fall back to the default values for these
variables.

To disable "debuild" sanitising these variables out of the environment
the "builder" attribute of the "debian/gbp.conf" is set to extend the
default "debuild -i -I" "git-buildpackage" uses to also preserve the
required environment variables.

Signed-off-by: Thomas Böhler <t.boehler@kunbus.com>