dts: Add Connect 5 with 2xCAN controller (MCP251863)
A 20MHz oscillator is used for OSC1 and OSC2 input. It is modled as
can_20mhz_osc fixed-clock. The SPI SCK must be must be less than or equal
to 0.85*(FSYSCLK/2). Thus a max spi freuenzy of 8.5 MHz is selected. In
h/w two different oscillators are used, one on each board. In the device
tree only one can_20mhz_osc is provided as the ossilators are identical
and a "fixed-clock" regulator is used.

The CAN termination for can0 can be switched on by driving gpio 16 of the
gpio expander on the core board high.

The CAN termination for can0 can be switched on by driving gpio 12 of the
gpio expander on the power board high.

The second CAN interface replaces the RS485. Thus uart4 (RS485) is
disabled.

Signed-off-by: Philipp Rosenberger <p.rosenberger@kunbus.com>