dts/connect5: PiBridge mux is active low
The default state on the PiBridge is backplane mode (eg. compatible to
older devices). Newer devices like RevPi Connect 5 can also switch the
pins to ethernet mode which supports gbit ethernet using all four
pairs. Therefore the multiplexer must ensure that the pins are
configured for PiBridge as default. Unfortunately this is not the case
with Connect 5 and only a pull-up resistor ensures, that the defaults
state is PiBridge mode.

Fix this and change the mux gpios to active low. Otherwise the piControl
driver would need to invert the logic which might break future designs,
which should handle the multiplexer default state correctly.

Fixes: 295d996609ff ("dts/connect5: add GPIOs for PiBridge gigabit ethernet")

Signed-off-by: Nicolai Buchwitz <n.buchwitz@kunbus.com>