can: dev: can_set_termination(): Allow gpio sleep
The current implementation of can_set_termination() sets the GPIO in a
context which cannot sleep. This is an issue if the GPIO controller can
sleep (e.g. since the concerning GPIO expander is connected via SPI or I2C).

Fix this by setting the termination resistor with gpiod_set_value_cansleep()
which instead of gpiod_set_value() allows it to sleep.

Signed-off-by: Nicolai Buchwitz <n.buchwitz@kunbus.com>