rtc: pcf2127: WORKAROUND: Force battery switch-over mode
With the new PCF2131 the battery switch-over mode is disable by default.
Any device using this RTC needs to enable battery switch-over manually.

As a workaround enable battery switch-over in standard mode and battery
low detection. This replicates the behavior of the older PCF2127/PCF2129
devices.

The RTC maintainer has started devleopment of a patch that would make it
possible to configure the battery switch-over behavior from user space:
https://lore.kernel.org/lkml/20241114085153e4e23a7f@mail.local/
When that patch is ready, this workaround should be replaced by this
patch and a proper configuration with user space tooling.

Signed-off-by: Philipp Rosenberger <p.rosenberger@kunbus.com>