pibridge-serdev: handle error frames
The current error handling in the io communication includes checks for errors
which may happen in response datagrams due to frame corruption on the wire.
However the RevPi IO modules may also encounter errors during the processing of
a request before a response is sent back.
For those cases message type 7 (PIBRIDGE_IO_CMD_ERR) is used in the response to
indicate an error.
Include a check for this message type into the error handling. This might be
extended in future when specific error values are sent with the data portion
of the datagram.

Signed-off-by: Lino Sanfilippo <l.sanfilippo@kunbus.com>