Revert "pibridge-serdev: handle error frames"
This reverts commit 2426f2738919fb4bc209efe2fbbde8ba2040fbb5.

As it turned out there is a bug in the firmware of AIO modules:
In the response frame sent for config messages (IOP_TYP1_CMD_DATA,
IOP_TYP1_CMD_DATA2, IOP_TYP1_CMD_DATA3) the 'cmd' field is set to
IOP_TYP1_CMD_ERR instead of the value sent in the request.
So with commit 2426f273891d each AIO response to a config request is
considered as an error frame and thus the error handling (i.e. deactivation
of the concerning module) is executed. So revert that change to avoid
the error processing.

Signed-off-by: Lino Sanfilippo <l.sanfilippo@kunbus.com>